requirements.txt
tensorflow-gpu==2.10.1
keras==2.10.0
numpy==1.24.1
pandas==1.5.3
matplotlib==3.6.3
opencv-python==4.7.0.68
boto3==1.26.69
botocore==1.29.69
tqdm==4.64.1
jupyter==1.0.0
attrdict3==2.0.2
PyYAML==6.0